fix(quark): reuse auth mapping for share-tree
